trying to clean up some of the code to make it more readable and / or portable
I have reservations about some of the calculations accuracies (not necessarily the steady states though) given that there are some very jagged plots in the 2d phase planes and im worried that we're losing resolution here. Can be solved by a larger # of steps but this could prove to be a performance hit. This also may be solved in time with a fast-slow subsystem analysis. To be brought up in the meeting later today